整合维特比自定义词典代码
